Fracking Tower And Bedrock Oil Deposit
AWESOME work by Ratainen

Added Fracking Tower.
Added Fracking Solution.
Added Dead Plants and all the other polluted stuff, because of the tower.
Added Bedrock Oil Deposit
Took some duplicate code and put it to TileEntityOilDrillBase. (that was how it was done in the 1.7.10 version and my brain doesn't explode when the code looks more similiar to that, so that's pretty much the real reason for that.)

The tower will pollute the nearby environment by killing plants and all that stuff.
Bedrock oil deposit has that dead stuff over it on the ground so you can find it easier.

by Alcater

making all that work as intended and fixing bugs